84% / 67% / 82%
Back to this because of... "git" issues

Mutation testing #56

+++ backend/omniplanner/src/main/java/com/main/omniplanner/user/EventService.java
@@ -39,6 +39,9 @@ public class EventService {
 
     public boolean deleteEvent(int id, String token) {
         Integer userId = userRepository.getIdByToken(token);
+        if (userId == null) {
+            throw new NullPointerException("Invalid token");
+        }
         return eventRepository.deleteEvent(id, userId) > 0;
     }
